--- src/main/java/com/proyect/tinder/enumerations/Sex.java
@@ -1,5 +1,0 @@
-package com.proyect.tinder.enumerations;
-
-public enum Sex {
-    MALE,FEMALE;
-}
