+++ src/main/java/com/proyect/tinder/enums/Sex.java
@@ -1,0 +1,5 @@
+package com.proyect.tinder.enums;
+
+public enum Sex {
+    MALE,FEMALE;
+}
